backend/requirements.txt
asyncio
nest-asyncio
asyncpg
sqlalchemy
python-dotenv
fastapi
uvicorn
jwt